--- a/package.json
+++ b/package.json
@@ -107,7 +107,7 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.19.3",
-    "@babel/core": "^7.20.2",
+    "@babel/core": "^7.23.2",
     "@babel/eslint-parser": "^7.19.1",
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-syntax-object-rest-spread": "^7.8.3",
@@ -124,7 +124,6 @@
     "acorn-jsx": "^5.3.2",
     "acorn-static-class-features": "^1.0.0",
     "archiver": "^5.3.1",
-    "babel-core": "^6.26.3",
     "babel-jest": "^29.7.0",
     "babel-loader": "^9.1.3",
     "chokidar": "^3.5.3",