GitHub Issue (if applicable): n/a

I optimized webpack config a bit (without changing output format) and also updated all related loaders. This cut webpack compilation time in ~half on my machine, which in pair with added `yarn watch:webpack` allows to iterate a lot quicker, without waiting for full build process after each change. 

<details>
<summary>Time changes on my machine</summary>

(With preinstalled dependencies)

Before optimizations:

```
❯ yarn build:development
...
✨  Done in 9.89s.

❯ yarn build:production
...
✨  Done in 13.69s.
```

After optimizations:

```
❯ yarn build:development
...
✨  Done in 6.40s.

❯ yarn build:production
...
✨  Done in 8.93s.
```

(Difference isn't 2x because I only optimized webpack part of pipeline, index generators were kept intact)
</details>